# swagger.json
{"swagger":"2.0","info":{"description":"Class Registration System - REST API","version":"1.0.0","title":"Class Registration System","contact":{"name":"Vineela Boddula","url":"https://www.linkedin.com/in/vineela-boddula-48152219/","email":"boddulavineela@gmail.com"},"license":{}},"host":"localhost:8080","basePath":"/","tags":[{"name":"Authentication","description":"Operations pertaining to authentication into Class Registration System"},{"name":"Class","description":"Operations pertaining to class in Class Registration System"},{"name":"Professor","description":"Operations pertaining to professor in Class Registration System"},{"name":"Student","description":"Operations pertaining to student in Class Registration System"}],"paths":{"/api/auth/signin":{"post":{"tags":["Authentication"],"summary":"authenticateUser","operationId":"authenticateUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"loginRequest","description":"loginRequest","required":true,"schema":{"$ref":"#/definitions/LoginRequest"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}},"security":[{"JWT":["global"]}],"deprecated":false}},"/api/auth/signup":{"post":{"tags":["Authentication"],"summary":"registerUser","operationId":"registerUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"signUpRequest","description":"signUpRequest","required":true,"schema":{"$ref":"#/definitions/SignUpRequest"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}}},"security":[{"JWT":["global"]}],"deprecated":false}},"/api/v1/classes":{"get":{"tags":["Class"],"summary":"View a list of available classes","operationId":"getAllClassesUsingGET","produces":["application/json"],"responses":{"200":{"description":"Successfully retrieved list","schema":{"type":"array","items":{"$ref":"#/definitions/ClassDTO"}}},"401":{"description":"You are not authorized to view the resource"},"403":{"description":"Accessing the resource you were trying to reach is forbidden"},"404":{"description":"The resource you were trying to reach is not found"}},"security":[{"JWT":["global"]}],"deprecated":false},"post":{"tags":["Class"],"summary":"Add a class","operationId":"createClassUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"newClass","description":"Class object store in database table","required":true,"schema":{"$ref":"#/definitions/Class"}}],"responses":{"201":{"description":"Created","schema":{"type":"object"}}},"security":[{"JWT":["global"]}],"deprecated":false}},"/api/v1/classes/{id}":{"get":{"tags":["Class"],"summary":"Get a class by Id","operationId":"getClassByIdUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Class id from which class object will retrieve","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ClassDTO"}}},"security":[{"JWT":["global"]}],"deprecated":false},"put":{"tags":["Class"],"summary":"Update a class","operationId":"updateClassUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Class Id to update class object","required":true,"type":"integer","format":"int64"},{"in":"body","name":"updateClass","description":"Update class object","required":true,"schema":{"$ref":"#/definitions/Class"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Class"}}},"security":[{"JWT":["global"]}],"deprecated":false},"delete":{"tags":["Class"],"summary":"Delete a class","operationId":"deleteClassUsingDELETE","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Class Id from which class object will delete from database table","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}},"security":[{"JWT":["global"]}],"deprecated":false}},"/api/v1/professors":{"get":{"tags":["Professor"],"summary":"View a list of available professors","operationId":"getAllProfessorsUsingGET","produces":["application/json"],"responses":{"200":{"description":"Successfully retrieved list","schema":{"type":"array","items":{"$ref":"#/definitions/ProfessorDTO"}}},"401":{"description":"You are not authorized to view the resource"},"403":{"description":"Accessing the resource you were trying to reach is forbidden"},"404":{"description":"The resource you were trying to reach is not found"}},"security":[{"JWT":["global"]}],"deprecated":false},"post":{"tags":["Professor"],"summary":"Add a professor","operationId":"createProfessorUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"newProfessor","description":"Professor object store in database table","required":true,"schema":{"$ref":"#/definitions/Professor"}}],"responses":{"201":{"description":"Created","schema":{"type":"object"}}},"security":[{"JWT":["global"]}],"deprecated":false}},"/api/v1/professors/{id}":{"get":{"tags":["Professor"],"summary":"Get a professor by Id","operationId":"getProfessorByIdUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Professor id from which professor object will retrieve","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ProfessorDTO"}}},"security":[{"JWT":["global"]}],"deprecated":false},"put":{"tags":["Professor"],"summary":"Update a professor","operationId":"updateProfessorUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Professor Id to update professor object","required":true,"type":"integer","format":"int64"},{"in":"body","name":"updateProfessor","description":"Update professor object","required":true,"schema":{"$ref":"#/definitions/Professor"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Professor"}}},"security":[{"JWT":["global"]}],"deprecated":false},"delete":{"tags":["Professor"],"summary":"Delete a professor","operationId":"deleteProfessorUsingDELETE","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Professor Id from which professor object will delete from database table","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}},"security":[{"JWT":["global"]}],"deprecated":false}},"/api/v1/professors/{id}/assignClass":{"put":{"tags":["Professor"],"summary":"Assign professor to a class","operationId":"assignClassUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Professor Id","required":true,"type":"integer","format":"int64"},{"in":"body","name":"theClass","description":"Class assigned by professor object","required":true,"schema":{"$ref":"#/definitions/Class"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Professor"}}},"security":[{"JWT":["global"]}],"deprecated":false}},"/api/v1/students":{"get":{"tags":["Student"],"summary":"View a list of available students","operationId":"getAllStudentsUsingGET","produces":["application/json"],"responses":{"200":{"description":"Successfully retrieved list","schema":{"type":"array","items":{"$ref":"#/definitions/StudentDTO"}}},"401":{"description":"You are not authorized to view the resource"},"403":{"description":"Accessing the resource you were trying to reach is forbidden"},"404":{"description":"The resource you were trying to reach is not found"}},"security":[{"JWT":["global"]}],"deprecated":false},"post":{"tags":["Student"],"summary":"Add a student","operationId":"createStudentUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"newStudent","description":"Student object store in database table","required":true,"schema":{"$ref":"#/definitions/Student"}}],"responses":{"201":{"description":"Created","schema":{"type":"object"}}},"security":[{"JWT":["global"]}],"deprecated":false}},"/api/v1/students/{id}":{"get":{"tags":["Student"],"summary":"Get a professor by Id","operationId":"getStudentByIdUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Student id from which professor object will retrieve","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/StudentDTO"}}},"security":[{"JWT":["global"]}],"deprecated":false},"put":{"tags":["Student"],"summary":"Update a student","operationId":"updateStudentUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Student Id to update student object","required":true,"type":"integer","format":"int64"},{"in":"body","name":"updateStudent","description":"Update student object","required":true,"schema":{"$ref":"#/definitions/Student"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Student"}}},"security":[{"JWT":["global"]}],"deprecated":false},"delete":{"tags":["Student"],"summary":"Delete a student","operationId":"deleteStudentUsingDELETE","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Student Id from which student object will delete from database table","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}},"security":[{"JWT":["global"]}],"deprecated":false}},"/api/v1/students/{id}/registerClass":{"put":{"tags":["Student"],"summary":"Register student to a class","operationId":"registerClassUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Student Id","required":true,"type":"integer","format":"int64"},{"in":"body","name":"theClass","description":"Class registered by student object","required":true,"schema":{"$ref":"#/definitions/Class"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Student"}}},"security":[{"JWT":["global"]}],"deprecated":false}}},"securityDefinitions":{"JWT":{"type":"apiKey","name":"Authorization","in":"header"}},"definitions":{"Class":{"type":"object","required":["className","classNumber"],"properties":{"classDescription":{"type":"string","description":"The Class Description"},"className":{"type":"string","description":"The Class Name"},"classNumber":{"type":"string","description":"The Class Number"},"id":{"type":"integer","format":"int64","description":"The database generated Class ID"}},"title":"Class","description":"All details about the Class. "},"ClassDTO":{"type":"object","required":["className","classNumber","id"],"properties":{"classDescription":{"type":"string","description":"The Class Description"},"className":{"type":"string","description":"The Class Name"},"classNumber":{"type":"string","description":"The Class Number"},"id":{"type":"integer","format":"int64","description":"The database generated Class ID"},"professorAssigned":{"type":"string","description":"Professors assigned to the class (i.e Full Name with semicolon separated)"},"studentsCount":{"type":"integer","format":"int32","description":"Number of students registered to the class"}},"title":"ClassDTO","description":"All details about the ClassDTO. "},"LoginRequest":{"type":"object","required":["password","username"],"properties":{"password":{"type":"string","description":"Password"},"username":{"type":"string","description":"User Name"}},"title":"LoginRequest","description":"All details about the Login. "},"Professor":{"type":"object","required":["emailAddress","firstName","lastName"],"properties":{"emailAddress":{"type":"string","description":"The Professor Email Address"},"firstName":{"type":"string","description":"The Professor First Name"},"id":{"type":"integer","format":"int64","description":"The database generated Professor ID"},"lastName":{"type":"string","description":"The Professor Last Name"},"officeAddress":{"type":"string","description":"The Professor Office Address"},"phoneNumber":{"type":"string","description":"The Professor Phone Number"}},"title":"Professor","description":"All details about the Professor. "},"ProfessorDTO":{"type":"object","required":["emailAddress","firstName","id","lastName"],"properties":{"classesTeaching":{"type":"string","description":"Classes Taught by the Professor (i.e Class Name with semicolon separated)"},"emailAddress":{"type":"string","description":"The Professor Email Address"},"firstName":{"type":"string","description":"The Professor First Name"},"id":{"type":"integer","format":"int64","description":"The database generated Professor ID"},"lastName":{"type":"string","description":"The Professor Last Name"},"officeAddress":{"type":"string","description":"The Professor Office Address"},"phoneNumber":{"type":"string","description":"The Professor Phone Number"}},"title":"ProfessorDTO","description":"All details about the ProfessorDTO. "},"SignUpRequest":{"type":"object","required":["email","name","password","username"],"properties":{"email":{"type":"string","description":"Email"},"name":{"type":"string","description":"Name"},"password":{"type":"string","description":"Password"},"role":{"type":"array","items":{"type":"string"}},"username":{"type":"string","description":"User Name"}},"title":"SignUpRequest","description":"All details about the Signup. "},"Student":{"type":"object","required":["firstName","lastName"],"properties":{"address":{"type":"string","description":"The Student Address"},"emailAddress":{"type":"string","description":"The Student Email Address"},"firstName":{"type":"string","description":"The Student First Name"},"id":{"type":"integer","format":"int64","description":"The database generated Student ID"},"lastName":{"type":"string","description":"The Student Last Name"},"phoneNumber":{"type":"string","description":"The Student Phone Number"}},"title":"Student","description":"All details about the Student. "},"StudentDTO":{"type":"object","required":["emailAddress","firstName","id","lastName"],"properties":{"address":{"type":"string","description":"The Student Address"},"classesRegistered":{"type":"string","description":"Classes Registered by the Student (i.e Class Name with semicolon separated)"},"emailAddress":{"type":"string","description":"The Student Email Address"},"firstName":{"type":"string","description":"The Student First Name"},"id":{"type":"integer","format":"int64","description":"The database generated Student ID"},"lastName":{"type":"string","description":"The Student Last Name"},"phoneNumber":{"type":"string","description":"The Student Phone Number"}},"title":"StudentDTO","description":"All details about the StudentDTO. "}}}